+++ CHANGELOG.md
@@ -15,5 +15,6 @@
 
 ### Added
 
+-   Google Drive elements list.
 -   AuthenticatorDrive class for Google Drive authentication.
 -   Google Drive dependencies to composer.json.